completa

+++ app/Http/Controllers/WineController.php
@@ -3,27 +3,31 @@
 namespace App\Http\Controllers;
 
 use Auth;
+use DB;
 use App\Models\Wine;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Http\Response;
+use App\Models\Like;
+use App\Models\Favourite;
 
 class WineController extends Controller
 {
     function __construct()
     {
-        
     }
+
     public function index()
     {
-        $wines = Wine::all();
-
+        $this->authorize('viewAny', Wine::class);
+        $wines = Wine::query()->paginate(10);
         return view('wine.index', ['wines' => $wines]);
     }
 
     public function create()
     {
+        $this->authorize('create', Wine::class);
         return view('wine.create');
     }
 
@@ -69,7 +73,7 @@ class WineController extends Controller
             $wine->image = $newImage;
         }
         $wine->save();
-        return redirect('/wines')->with('status','Producto Añadido');
+        return redirect('/wines')->with('status', 'Producto Añadido');
     }
 
     public function show(Wine $wine)
@@ -79,6 +83,7 @@ class WineController extends Controller
 
     public function edit(Wine $wine)
     {
+        $this->authorize('update', $wine);
         return view('wine.edit', ['wine' => $wine]);
     }
 
@@ -100,21 +105,96 @@ class WineController extends Controller
             $newImage = time() . $image->getClientOriginalName(); //llamar a la imagen con el nombre del archivo que sube el usuario para su avatar
             Storage::disk('wines')->put($newImage, File::get($image)); // Guardar la imagen en la carpeta (storage/app/users)
             $wine->image = $newImage;
-        }    
+        }
         $wine->save();
-        return redirect('/wines')->with('status','Producto Actualizado');
+        return redirect('/wines')->with('status', 'Producto Actualizado');
     }
 
     public function destroy($id)
     {
-
+        $user = Auth::user();
+        $wine = Wine::find($id);
+        $this->authorize('delete', Wine::class);
+        Storage::disk('wines')->delete($wine->image);/*Borramos la imagen del vino del disco*/
+        Like::where('wine_id', $id)->delete();/*Borramos sus likes*/
+        Favourite::where('wine_id', $id)->delete();/*Borramos sus comentarios*/
         Wine::destroy($id);
-        return back()->with('status','Producto Borrado');
+        return back()->with('status', 'Producto Borrado');
     }
 
     public function getImage($filename)
     {
         $file = Storage::disk('wines')->get($filename);
         return new Response($file, 200);
+    }
+    public function tintos(Request $request)
+    {
+        $name = $request->name;
+        $origin = $request->origin;
+        $type = $request->type;
+        $query = Wine::where('category', '=', 'Tinto'); //Filtrar todos los que su categoria sea Tinto
+        if (!empty($name)) {
+            $query = $query->where('name', 'like', "%$name%");
+        }
+        if (!empty($origin)) {
+            $query = $query->where('origin', 'like', "%$origin%");
+        }
+        if (!empty($type)) {
+            $query = $query->where('type', 'like', "%$type%");
+        }
+        $wines = $query->get(); //Recogemos los resultados del filtrado
+        $wines = $query->paginate(9);
+        return view('wine.tintos', ['wines' => $wines, 'name' => $name, 'origin' => $origin, 'type' => $type]);
+    }
+    public function rosados(Request $request)
+    {
+        $name = $request->name;
+        $origin = $request->origin;
+        $type = $request->type;
+        $query = Wine::where('category', '=', 'Rosado'); //Filtrar todos los que su categoria sea Rosado
+        if (!empty($name)) {
+            $query = $query->where('name', 'like', "%$name%");
+        }
+        if (!empty($origin)) {
+            $query = $query->where('origin', 'like', "%$origin%");
+        }
+        if (!empty($type)) {
+            $query = $query->where('type', 'like', "%$type%");
+        }
+        $wines = $query->get(); //Recogemos los resultados del filtrado
+        $wines = $query->paginate(9);
+        return view('wine.rosados', ['wines' => $wines, 'name' => $name, 'origin' => $origin, 'type' => $type]);
+    }
+    public function blancos(Request $request)
+    {
+        $name = $request->name;
+        $origin = $request->origin;
+        $type = $request->type;
+        $query = Wine::where('category', '=', 'Blanco'); //Filtrar todos los que su categoria sea Blanco
+        if (!empty($name)) {
+            $query = $query->where('name', 'like', "%$name%");
+        }
+        if (!empty($origin)) {
+            $query = $query->where('origin', 'like', "%$origin%");
+        }
+        if (!empty($type)) {
+            $query = $query->where('type', 'like', "%$type%");
+        }
+        $wines = $query->get(); //Recogemos los resultados del filtrado
+        $wines = $query->paginate(9);
+        return view('wine.blancos', ['wines' => $wines, 'name' => $name, 'origin' => $origin, 'type' => $type]);
+    }
+    public function topUsers()
+    {
+        //Consulta para obtener y ordenar por mayor numero de likes dados por los usuarios y los obtengo con valor wine_id
+        $topwines = Wine::select(DB::raw('wine_id'))
+            ->leftjoin('likes', 'wine_id', '=', 'wine_id')
+            ->groupBy('wine_id')
+            ->orderBy(DB::raw('COUNT(wine_id)'), 'DESC')
+            ->take(9)->get();//obtiene solo 9,para mostrar en pantalla unicamente los 9 más votados
+
+        $wines = Wine::all();
+
+        return view('like.topusers', ['topwines' => $topwines, 'wines' => $wines]);
     }
 }
